refactor content-process for sfari - 4
